removed ayto kick and terms counter and change the style of result message to fit with odin mod message style
